update : relocate UserDisplayModel
