+++ base/data/models/UserDisplayModel.ts
@@ -1,0 +1,7 @@
+export interface UserDisplayModel {
+    id: string;
+    name: string;
+    avatar: string;
+    username: string;
+    desc?:string;
+  }
